requirements.txt
requests==2.22.0
ratelimit==2.2.1
openai
tqdm
python-dotenv
